--- a/docker/aws-s3-core/Dockerfile.ubuntu
+++ b/docker/aws-s3-core/Dockerfile.ubuntu
@@ -18,6 +18,8 @@ RUN apt-get -y update && apt-get install --no-install-recommends -y \
 RUN mkdir /root/build
 WORKDIR /root/build
 
+COPY docker/utils/get_make_options.sh .
+
 # aws s3/core build and install
 RUN git clone https://github.com/aws/aws-sdk-cpp.git
 WORKDIR /root/build/aws-sdk-cpp
@@ -25,7 +27,7 @@ RUN git checkout tags/${aws_release} -b ${aws_release}
 # -DCUSTOM_MEMORY_MANAGEMENT=0 is added to avoid Aws::String and std::string issue
 # -DENABLE_TESTING=OFF for a weird failing test HttpClientTest.TestRandomURLWithProxyAndOtherDeclaredAsNonProxyHost
 RUN cmake . -DBUILD_ONLY="s3;core" -DCMAKE_BUILD_TYPE=RelWithDebInfo -DBUILD_SHARED_LIBS=OFF -DCUSTOM_MEMORY_MANAGEMENT=0 -DENABLE_TESTING=OFF
-RUN make && make install
+RUN . /root/build/get_make_options.sh && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD && make install -j $MAKE_JOBS -l $MAKE_MAX_LOAD
 
 FROM ubuntu:${os_release}
 COPY --from=builder /usr/local/include/. /usr/local/include/.