--- a/docker/emp/Dockerfile.ubuntu
+++ b/docker/emp/Dockerfile.ubuntu
@@ -19,6 +19,7 @@ WORKDIR /root/build
 
 # Build and install emp packages
 FROM dev-environment AS emp-builder
+COPY docker/utils/get_make_options.sh .
 ARG emp_tool_release="0.2.3"
 ARG emp_release="0.2.2"
 ARG DEBIAN_FRONTEND=noninteractive
@@ -32,21 +33,21 @@ RUN git clone -b${emp_tool_release} https://github.com/emp-toolkit/emp-tool.git
 WORKDIR /root/build/emp-tool
 RUN sed -i "s/SHARED/STATIC/" CMakeLists.txt
 RUN cmake -DBUILD_SHARED_LIBS=OFF -DTHREADING=ON .
-RUN make && make install
+RUN . /root/build/get_make_options.sh && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD install
 
 # emp-ot build and install
 WORKDIR /root/build
 RUN git clone -b${emp_release} https://github.com/emp-toolkit/emp-ot.git
 WORKDIR /root/build/emp-ot
 RUN cmake -DBUILD_SHARED_LIBS=OFF -DTHREADING=ON .
-RUN make && make install
+RUN . /root/build/get_make_options.sh && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD install
 
 # emp-sh2pc build and install
 WORKDIR /root/build
 RUN git clone -b${emp_release} https://github.com/emp-toolkit/emp-sh2pc.git
 WORKDIR /root/build/emp-sh2pc
 RUN cmake -DBUILD_SHARED_LIBS=OFF -DTHREADING=ON .
-RUN make && make install
+RUN . /root/build/get_make_options.sh && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD && make -j $MAKE_JOBS -l $MAKE_MAX_LOAD install
 
 FROM ubuntu:${os_release}
 RUN mkdir /usr/local/cmake