Summary:
# Context
This change is intended to help speed up the builds for FBPCF and all the dependent images. Right now we are using a C4.4xlarge instance for building the images, but we never use anywhere near the CPU or Memory limits for an instance that large. This change allows us to take more advantage of that instance size and improve our build times. Some examples when building with the FBPCS machine are that we were only using about 6% CPU for a build and this change increased it to ~25% and reduced the build time from ~40 minutes to ~19 minutes.
https://pxl.cl/2qpD2 (https://github.com/facebookresearch/fbpcf/commit/7566956549c53453c504db141be371b74d6b1165)
https://pxl.cl/2qpD1

# This Diff
This diff updates the FBPCF builds to take advantage of the same build features. Right now builds are taking ~ 55 mintues. After the change, they are taking less than 30 minutes.

## Before the change:
```
build binaries completed successfully

  0.04s user 0.05s system 0% cpu 56:40.53 total
```
Logs: https://www.internalfb.com/intern/paste/P603909791/

##After the change:
```
build binaries completed successfully

  0.01s user 0.04s system 0% cpu 29:26.80 total
```
Logs: https://www.internalfb.com/intern/paste/P603946390/

Differential Revision: D42626641